// src/generated/models/UpdateUserRequest.generated.ts
/* eslint-disable */
/**
 * This file was automatically generated.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file.
 *
 * =>  @see ./src/schemas/models/UpdateUserRequest.yml
 *
 * and run "npm run schema-to-interface" or "npm run generate" to regenerate this file.
 */

import Ajv, {ValidateFunction} from "ajv";
import addFormats from "ajv-formats";

export const schemaForUpdateUserRequest = {
  $id: "UpdateUserRequest.yml",
  type: "object",
  properties: {email: {type: "string"}, firstName: {type: "string"}, lastName: {type: "string"}}
};

export function validateUpdateUserRequest(o: object): {isValid: boolean; validate: ValidateFunction} {
  const ajv = new Ajv();
  addFormats(ajv);
  ajv.addKeyword("example");

  const validate = ajv.compile(schemaForUpdateUserRequest);
  return {isValid: validate(o), validate: validate};
}

export interface UpdateUserRequest {
  email?: string;
  firstName?: string;
  lastName?: string;
}